refactor: remove sparque as an extension and use it as a default feature flag from Eisi96
