+++ src/app/core/core.module.ts
@@ -14,6 +14,7 @@ import { MockInterceptor } from './interceptors/mock.interceptor';
 import { PaymentPayoneInterceptor } from './interceptors/payment-payone.interceptor';
 import { PreviewInterceptor } from './interceptors/preview.interceptor';
 import { InternationalizationModule } from './internationalization.module';
+import { ServicesModule } from './services.module';
 import { StateManagementModule } from './state-management.module';
 import { DefaultErrorHandler } from './utils/default-error-handler';
 
@@ -26,6 +27,7 @@ import { DefaultErrorHandler } from './utils/default-error-handler';
     IdentityProviderModule,
     InternationalizationModule,
     ServiceWorkerModule.register('ngsw-worker.js', { enabled: SERVICE_WORKER }),
+    ServicesModule,
     StateManagementModule,
   ],
   providers: [
