+++ src/app/core/models/sparque-config/sparque-config.model.ts
