+++ src/app/core/utils/types.ts
@@ -1,0 +1,4 @@
+/**
+ * generic type to extract return type from a property K out of T
+ */
+export type ReturnTypeFromKey<T, K extends keyof T> = T[K];
