--- src/app/extensions/sparque/store/sparque-config/index.ts
@@ -1,3 +1,0 @@
-// API to access ngrx sparqueConfig state
-export * from './sparque-config.actions';
-export * from './sparque-config.selectors';
