--- src/app/extensions/sparque/store/sparque-config/sparque-config.selectors.ts
@@ -1,7 +1,0 @@
-import { createSelector } from '@ngrx/store';
-
-import { getSparqueState } from '../sparque-store';
-
-export const getSparqueConfig = createSelector(getSparqueState, state => state?.sparqueConfig);
-
-export const getSparqueConfigEndpoint = createSelector(getSparqueConfig, config => config?.endPoint);
