--- src/app/extensions/sparque/store/sparque-store.ts
@@ -1,8 +1,0 @@
-import { createFeatureSelector } from '@ngrx/store';
-
-import { SparqueConfig } from '../models/sparque-config/sparque-config.model';
-
-export interface SparqueState {
-  sparqueConfig: SparqueConfig;
-}
-export const getSparqueState = createFeatureSelector<SparqueState>('sparque');
