+++ src/app/shared/shared.module.ts
@@ -34,7 +34,6 @@ import { QuickorderExportsModule } from '../extensions/quickorder/exports/quicko
 import { QuotingExportsModule } from '../extensions/quoting/exports/quoting-exports.module';
 import { RatingExportsModule } from '../extensions/rating/exports/rating-exports.module';
 import { RecentlyExportsModule } from '../extensions/recently/exports/recently-exports.module';
-import { SparqueExportsModule } from '../extensions/sparque/exports/sparque-exports.module';
 import { StoreLocatorExportsModule } from '../extensions/store-locator/exports/store-locator-exports.module';
 import { TactonExportsModule } from '../extensions/tacton/exports/tacton-exports.module';
 import { WishlistsExportsModule } from '../extensions/wishlists/exports/wishlists-exports.module';
@@ -181,7 +180,6 @@ const importExportModules = [
   RecentlyExportsModule,
   RoleToggleModule,
   RouterModule,
-  SparqueExportsModule,
   StoreLocatorExportsModule,
   SwiperModule,
   TactonExportsModule,
