+++ src/environments/environment.model.ts
@@ -5,7 +5,7 @@ import { DataRetentionPolicy } from 'ish-core/utils/meta-reducers';
 import { MultiSiteLocaleMap } from 'ish-core/utils/multi-site/multi-site.service';
 
 import { AddressDoctorConfig } from '../app/extensions/address-doctor/models/address-doctor/address-doctor-config.model';
-import { SparqueConfig } from '../app/extensions/sparque/models/sparque-config/sparque-config.model';
+import { SparqueConfig } from '../app/core/models/sparque-config/sparque-config.model';
 import { TactonConfig } from '../app/extensions/tacton/models/tacton-config/tacton-config.model';
 
 export interface Environment {
